Move TransferPlanner to its own file
