+++ src/Pi.cpp
@@ -51,8 +51,8 @@
 #include "Star.h"
 #include "StringF.h"
 #include "SystemInfoView.h"
-#include "SystemView.h"
 #include "Tombstone.h"
+#include "TransferPlanner.h"
 #include "WorldView.h"
 #include "galaxy/GalaxyGenerator.h"
 #include "libs.h"
