+++ src/SystemView.h
@@ -8,6 +8,7 @@
 #include "DeleteEmitter.h"
 #include "Frame.h"
 #include "Input.h"
+#include "TransferPlanner.h"
 #include "enum_table.h"
 #include "graphics/Drawables.h"
 #include "matrix4x4.h"
@@ -20,12 +21,6 @@ class Orbit;
 class Ship;
 class Game;
 class Body;
-
-enum BurnDirection {
-	PROGRADE,
-	NORMAL,
-	RADIAL,
-};
 
 enum ShipDrawing {
 	BOXES,
@@ -43,37 +38,6 @@ enum ShowLagrange {
 	LAG_ICON,
 	LAG_ICONTEXT,
 	LAG_OFF
-};
-
-class TransferPlanner {
-public:
-	TransferPlanner();
-	vector3d GetVel() const;
-	vector3d GetOffsetVel() const;
-	vector3d GetPosition() const;
-	double GetStartTime() const;
-	void SetPosition(const vector3d &position);
-	void IncreaseFactor(), ResetFactor(), DecreaseFactor();
-	void AddStartTime(double timeStep);
-	void ResetStartTime();
-	double GetFactor() const { return m_factor; }
-	void AddDv(BurnDirection d, double dv);
-	double GetDv(BurnDirection d);
-	void ResetDv(BurnDirection d);
-	void ResetDv();
-	std::string printDeltaTime();
-	std::string printDv(BurnDirection d);
-	std::string printFactor();
-
-private:
-	double m_dvPrograde;
-	double m_dvNormal;
-	double m_dvRadial;
-	double m_factor;				   // dv multiplier
-	const double m_factorFactor = 5.0; // m_factor multiplier
-	vector3d m_position;
-	vector3d m_velocity;
-	double m_startTime;
 };
 
 struct Projectable {
